+++ lib/src/formatters/formatters.dart
@@ -1,2 +1,3 @@
 export 'lib_phone_as_you_type_formatter.dart';
+export 'phone_number_digit_length_limiting_formatter.dart';
 export 'phone_number_digits_only_formatter.dart';